Wait for all hiscore results before updating the UI
This commit merges the futures returned from the hiscore client so that
the plugin waits for all hiscore results to come back before updating
anything in the panel. This looks much smoother than the current
approach where each player's panel updates as soon as that player's
hiscores are retrieved, which can look choppy. All hiscores are
retrieved in parallel so no extra performance hit is incurred by these
changes.