Avoid auto-updating the group before the widgets load
In cases where the group iron tab is selected in the grouping menu but
the menu has not been open since the player last logged in, the widgets
that provide the group name and group members are present but do not
have the data loaded into them. The data is only present after the menu
has been opened at least once.

The plugin was incorrectly using this empty data to auto-detect that the
group had changed, which overwrote the group plugin config with junk
data and broke the plugin. The plugin now more strictly checks that the
group info coming from the widgets is present and valid before
overwriting the existing configuration.